+++ Src-imro/IMROTbl_T3020base.cpp
@@ -553,7 +553,7 @@ void IMROTbl_T3020base::edit_ROI2tbl( tconstImroROIs vR, const IMRO_Attr &A )
     if( joinTips )
         refIdx = 2;
 
-    for( int ib = 0, nb = vR.size(); ib < nb; ++ib ) {
+    for( int ib = 0, nb = (int)vR.size(); ib < nb; ++ib ) {
 
         const IMRO_ROI  &B = vR[ib];
